requirements.txt
absl-py==2.2.0
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
alabaster==1.0.0
altair==4.2.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
audio-recorder-streamlit==0.0.8
babel==2.17.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
Django==5.1.7
docutils==0.21.2
entrypoints==0.4
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==1.34.1
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.49.0rc1
gTTS==2.5.1
h11==0.14.0
h5py==3.13.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imagesize==1.4.1
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==3.9.0
kiwisolver==1.4.8
libclang==18.1.1
libsql-client==0.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.1.0
namex==0.0.8
nltk==3.9.1
numpy==1.26.4
openai==1.35.3
opt_einsum==3.4.0
optree==0.14.1
packaging==23.2
pandas==1.5.3
pillow==10.3.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==3.20.3
py4j==0.10.9.5
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
Pympler==1.0.1
pyparsing==3.2.1
pyspark==3.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.7.1
roman-numerals-py==3.1.0
rpds-py==0.18.1
rsa==4.9
scikit-fuzzy==0.5.0
semver==3.0.2
setuptools==77.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
SpeechRecognition==3.14.2
Sphinx==8.2.3
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
streamlit==1.43.2
streamlit-ace==0.1.1
streamlit-chat==0.1.1
tenacity==8.4.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
validators==0.28.3
watchdog==4.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.19.2
